Splitted all modules related functions from utils into their own source file. Fixed some doc.
