+++ inc/bus.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <systemd/sd-bus.h>
-
 #include "utils.h"
 
 /*
