+++ inc/modules.h
@@ -1,0 +1,8 @@
+#include "log.h"
+
+void init_modules(const enum modules module);
+void init_module(int fd, enum modules module, void (*cb)(void));
+void set_self_deps(struct self_t *self);
+void disable_module(const enum modules module);
+void poll_cb(const enum modules module);
+void destroy_modules(const enum modules module);
